#36 #35 download gmic archive if needed

+++ build_tools.bash
@@ -52,7 +52,9 @@ function 1_clean_and_regrab_gmic_src () {
     mkdir src -p
     $PIP3 install -r dev-requirements.txt
     $PYTHON3 setup.py clean --all
-    wget ${GMIC_URL} -P src/ --no-check-certificate || { echo "Fatal gmic src archive download error" ; exit 1; }
+    if ! [ -f "$GMIC_ARCHIVE_GLOB"]; then
+        wget ${GMIC_URL} -P src/ --no-check-certificate || { echo "Fatal gmic src archive download error" ; exit 1; }
+    fi
     tar xzvf src/${GMIC_ARCHIVE_GLOB} -C src/ || { echo "Fatal gmic src archive extracting error" ; exit 1; }
     # Keep only gmic source's src directory
     mv src/gmic*/ src/gmic
